extract enrichment skeleton_augmentation and special_move_examples from ConditionExampleGenerator
